RDKAAMP-506 [AAMP] refactoring - change returned httpCode from long to int
Reason for change: code simplification
Test Procedure: regression only
Risks: low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_23_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_23_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_23_1

+++ MediaStreamContext.cpp
@@ -484,7 +484,7 @@ int MediaStreamContext::GetDefaultDurationBetweenPlaylistUpdates()
 /**
  * @brief ProcessPlaylist after MPD refresh
  */
-void MediaStreamContext::ProcessPlaylist(GrowableBuffer& newPlaylist, long http_error)
+void MediaStreamContext::ProcessPlaylist(GrowableBuffer& newPlaylist, int http_error)
 {
     context->ProcessPlaylist(newPlaylist, http_error);
 }
